# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Sierpe API
  description: |
    Self-hosted Stellar contract indexer. The events surface is a compatible
    superset of the proposed getEvents v2 RPC semantics (stellar#1872):
    positional topic filters, opaque cursors that encode the entire query,
    and scan-status honesty. Coverage is declared in every events response —
    an empty page always says whether it means "nothing happened" or "not
    indexed yet".
  version: 0.1.0-m1
  license:
    name: Apache-2.0
servers:
  - url: http://localhost:8080
...
paths:
...
  /v1/contracts/{id}/movements:
    get:
      tags: [public]
      summary: Token movements this contract took part in
      description: |
        Token transfers naming this contract as sender or recipient,
        whichever contract emitted them. This is the resource that answers
        "what came into and went out of my contract": a payment to a
        contract is emitted by the ASSET's SAC, so it appears here without
        the operator having to register that asset. Requires the
        `movements` kind on the registration.

        Not a balance. Value can also arrive without any SEP-41 transfer
        event (custom tokens, contracts crediting internal balances), and
        amounts are raw base units of different tokens — never sum across
        `tokenContractId`.
      parameters:
        - { $ref: "#/components/parameters/contractID" }
        - name: role
          in: query
          schema: { type: string, enum: [recipient, sender] }
          description: Omit for both directions.
        - name: token
          in: query
          schema: { type: string }
          description: >
            Emitting token contract (C... strkey). The asset's identity is
            its contract, never its SEP-0011 string, which is
            attacker-choosable for custom tokens.
        - name: type
          in: query
          schema: { type: string, enum: [transfer, mint, burn, clawback] }
        - name: startLedger
          in: query
          schema: { type: integer, minimum: 1 }
        - name: endLedger
          in: query
          schema: { type: integer, minimum: 1 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 1000, default: 100 }
        - name: cursor
          in: query
          schema: { type: string }
      responses:
        "200":
          description: One page of movements with scan status and coverage.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MovementsPage" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: The ADMIN_TOKEN boot configuration value.
    basicAuth:
      type: http
      scheme: basic
      description: >
        Only when the instance sets HTTP_BASIC_AUTH: then EVERY path in
        this document except /health and /ready additionally requires
        these credentials. Instances on private networking leave it unset
        and reads stay open.
  parameters:
    contractID:
      name: id
      in: path
      required: true
      schema: { type: string }
      description: C... strkey contract address.
  responses:
    BadRequest:
      description: Invalid input; the error message says exactly what to fix.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    NotFound:
      description: The contract is not registered on this instance.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
...
    Coverage:
      type: object
      description: >
        The range this instance can vouch for, FOR ONE KIND (rule 7).
        Coverage is a property of (contract, kind): a walk derives only the
        kinds the registration carried while it ran, so a kind added later
        is not covered below the anchor until the walk is reopened.
      properties:
        kind:
          type: string
          enum: [events, state, transfers, trustlines, movements]
          description: The data kind this declaration is about.
        kindDerived:
          type: boolean
          description: >
            False when the registration does not derive this kind at all —
            an empty page then means "not indexed", never "nothing
            happened", and indexedFrom/To are 0.
        requestedFromLedger: { type: integer }
        requestedToLedger:
          type: integer
          nullable: true
          description: null means the query follows the tip.
        indexedFromLedger:
          type: integer
          description: >
            Oldest ledger vouched for. Right after a registration this can
            be GREATER than indexedToLedger: the backfill is anchored a
            little past the live cursor so the ledgers closing while the
            ingesting process reloads its registry are walked by somebody.
            An inverted pair is not an error, it is an EMPTY window — the
            instance vouches for nothing yet — and it closes on its own
            within a reload interval. Do not compute a span from these two
            numbers without checking the order first.
        indexedToLedger: { type: integer }
        backfillPending:
          type: boolean
          description: The lower bound is still improving.
        clampedAt:
          type: integer
          nullable: true
          description: >
            History below this ledger is unreachable from current sources
            (persisted as a gap; the archive leg can heal it).
...
    Movement:
      type: object
      properties:
        id:
          type: string
          description: >
            The source transfer's id, which is also its event id. A
            self-transfer produces two rows sharing it, so key on (id, role).
        role: { type: string, enum: [recipient, sender] }
        tokenContractId:
          type: string
          description: The emitting token contract — the asset's identity.
        transferType: { type: string, enum: [transfer, mint, burn, clawback] }
        counterparty:
          type: string
          description: The other side; absent on mint (in) and burn (out).
        amount: { type: string, description: Exact i128 amount in raw token units. }
        rawXdr:
          type: string
          description: >
            Base64 XDR of the full ContractEvent this movement was decoded
            from — the original bytes, since the emitting token is usually
            not registered and no events row exists to join to. Absent only
            on rows ingested before migration 0011.
        txHash:
          type: string
          description: >
            Hex hash of the transaction the movement's event was emitted
            in. Absent only on rows derived before migration 0013 that the
            tx-hash reconciliation has not reached yet.
        ledger: { type: integer }
        ledgerClosedAt: { type: string, format: date-time }
    MovementsPage:
      type: object
      properties:
        movements:
          type: array
          items: { $ref: "#/components/schemas/Movement" }
        cursor: { type: string }
        scanStatus:
          type: string
          enum: [HAS_MORE, WAITING_FOR_LEDGERS, OLDEST_REACHED, COMPLETE]
        coverage: { $ref: "#/components/schemas/Coverage" }
        latestLedger: { type: integer }
        note:
          type: string
          description: States what this resource is not — read it before treating totals as balances.
...